+++ test/lib/ex_aws/lambda_test.exs
@@ -2,7 +2,7 @@ defmodule ExAws.LambdaTest do
   use ExUnit.Case, async: true
 
   test "#list_functions" do
-    assert {:ok, %{"Functions" => _}} = ExAws.Lambda.list_functions
+    assert {:ok, %{"Functions" => _}} = ExAws.Lambda.list_functions |> ExAws.request
   end
 
 end
